fix d1-blog-server example
